refactor(components): custom editor 컴포넌트 atoms로 이동
